@@ -1,42 +1,27 @@
 plugins {
     id 'com.android.library'
     id 'kotlin-android'
-    id("com.apollographql.apollo").version("2.5.9")
+    id("com.apollographql.apollo3").version("3.0.0-alpha03")
     id 'kotlin-kapt'
     id 'dagger.hilt.android.plugin'
 }
 
-kapt {
-    correctErrorTypes = true
-}
-
 apollo {
-    generateKotlinModels.set(true)
-    customTypeMapping = [
-            "ItemT": "com.austinhodak.tarkovapi.models.Item"
-    ]
+    packageName.set("com.austinhodak.tarkovapi")
 }
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 31
     buildToolsVersion "30.0.3"
 
     defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 30
+        minSdkVersion 23
+        targetSdkVersion 31
         versionCode 1
         versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles "consumer-rules.pro"
-
-        javaCompileOptions {
-            annotationProcessorOptions {
-                arguments += [
-                        "room.schemaLocation": "$projectDir/schemas".toString()
-                ]
-            }
-        }
     }
 
     buildTypes {
@@ -45,46 +30,46 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
     }
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+
     kotlinOptions {
         jvmTarget = '1.8'
+    }
+
+    kapt {
+        correctErrorTypes = true
     }
 
 }
 
 dependencies {
 
+    implementation("com.apollographql.apollo3:apollo-runtime:3.0.0-alpha03")
+    implementation("com.apollographql.apollo3:apollo-api:3.0.0-alpha03")
+
     implementation "androidx.core:core-ktx:${versions.ktx}"
-    implementation 'androidx.appcompat:appcompat:1.3.0'
-    implementation 'com.google.android.material:material:1.3.0'
+    implementation 'androidx.appcompat:appcompat:1.3.1'
+    implementation 'com.google.android.material:material:1.4.0'
+    implementation 'androidx.compose.ui:ui-graphics:1.0.2'
+
     testImplementation 'junit:junit:4.13.2'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.2'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.3'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
 
-    //Apollo
-    // The core runtime dependencies
-    implementation("com.apollographql.apollo:apollo-runtime:2.5.9")
-    // Coroutines extensions for easier asynchronicity handling
-    implementation("com.apollographql.apollo:apollo-coroutines-support:2.5.9")
-
-    implementation "com.google.dagger:hilt-android:2.37"
-    kapt "com.google.dagger:hilt-android-compiler:2.37"
-
-    implementation("com.apollographql.apollo:apollo-http-cache:2.5.9")
-    implementation("com.apollographql.apollo:apollo-normalized-cache-sqlite:2.5.9")
+    implementation "com.google.dagger:hilt-android:2.38.1"
+    kapt "com.google.dagger:hilt-android-compiler:2.38.1"
 
     implementation("androidx.room:room-runtime:${versions.room}")
     kapt "androidx.room:room-compiler:${versions.room}"
 
     implementation("androidx.room:room-ktx:${versions.room}")
 
-    implementation 'com.google.code.gson:gson:2.8.7'
+    implementation 'com.google.code.gson:gson:2.8.8'
 
-    implementation 'com.jakewharton.timber:timber:4.7.1'
-
-    implementation 'com.mikepenz:fastadapter:5.4.1'
+    implementation 'com.jakewharton.timber:timber:5.0.1'
 
 }
