@@ -1,11 +1,6 @@
 package com.austinhodak.thehideout.calculator.models
 
-import android.content.res.ColorStateList
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import com.austinhodak.thehideout.R
-import com.austinhodak.thehideout.databinding.ItemPickerCalculatorCharacterBinding
-import com.mikepenz.fastadapter.binding.AbstractBindingItem
+import java.io.Serializable
 
 data class Character(
     val name: String,
@@ -13,37 +8,7 @@ data class Character(
     val image: String,
     val c_type: String,
     val spawn_chance: Int? = null
-) : AbstractBindingItem<ItemPickerCalculatorCharacterBinding>() {
-
-    override val type: Int
-        get() = R.id.fast_adapter_calc_character
-
-    override fun createBinding(inflater: LayoutInflater, parent: ViewGroup?): ItemPickerCalculatorCharacterBinding {
-        return ItemPickerCalculatorCharacterBinding.inflate(inflater, parent, false)
-    }
-
-    override fun bindView(binding: ItemPickerCalculatorCharacterBinding, payloads: List<Any>) {
-        binding.character = this
-
-        when (c_type) {
-            "boss" -> {
-                binding.calcPickerCharacterIcon.setImageResource(R.drawable.icons8_crown_96)
-                binding.calcPickerCharacterIcon.imageTintList = ColorStateList.valueOf(binding.calcPickerCharacterIcon.context.resources.getColor(R.color.md_amber_500))
-            }
-            "raider",
-            "guard" -> {
-                binding.calcPickerCharacterIcon.setImageResource(R.drawable.icons8_defense_96)
-                binding.calcPickerCharacterIcon.imageTintList = ColorStateList.valueOf(binding.calcPickerCharacterIcon.context.resources.getColor(R.color.md_brown_500))
-            }
-            "player" -> {
-                binding.calcPickerCharacterIcon.setImageResource(R.drawable.icons8_army_star_96)
-                binding.calcPickerCharacterIcon.imageTintList = ColorStateList.valueOf(binding.calcPickerCharacterIcon.context.resources.getColor(R.color.md_grey_400))
-            }
-            else -> {
-                binding.calcPickerCharacterIcon.setImageDrawable(null)
-            }
-        }
-    }
+) : Serializable {
 
     data class Health(
         val head: Int,
@@ -51,9 +16,11 @@ data class Character(
         val stomach: Int,
         val arms: Int,
         val legs: Int
-    ) {
+    ) : Serializable {
+
         override fun toString(): String {
             return (head + thorax + stomach + (arms * 2) + (legs * 2)).toString()
         }
+
     }
 }
