@@ -1,20 +1,27 @@
 package com.austinhodak.tarkovapi.room.dao
 
-import androidx.room.Dao
-import androidx.room.Query
+import androidx.room.*
 import com.austinhodak.tarkovapi.room.models.Ammo
+import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface AmmoDao {
 
+    @Transaction
     @Query("SELECT * FROM ammo")
-    fun getAllAmmo(): List<Ammo>
+    fun getAllAmmo(): Flow<List<Ammo>>
 
-    @Query("SELECT * FROM ammo WHERE id = :id")
-    fun getAmmoByID(id: String): Ammo
+    @Transaction
+    @Query("SELECT * FROM ammo WHERE id = :id ORDER BY id DESC")
+    fun getAmmo(id: String): Flow<Ammo>
 
     @Query("SELECT * FROM ammo WHERE Caliber = :caliber")
-    fun getAmmoByCaliber(caliber: String): List<Ammo>
+    fun getAmmoByCaliber(caliber: String): Flow<List<Ammo>>
 
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insertAll(data: List<Ammo>)
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insert(data: Ammo)
 
 }
