@@ -1,20 +1,20 @@
 package com.austinhodak.tarkovapi.room.dao
 
-import androidx.lifecycle.LiveData
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import com.austinhodak.tarkovapi.room.models.Craft
+import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface CraftDao {
 
     @Query("SELECT * FROM crafts")
-    fun getAllCrafts(): List<Craft>
+    fun getAllCrafts(): Flow<List<Craft>>
 
     @Query("SELECT * FROM crafts WHERE requiredItems LIKE :id OR rewardItems LIKE :id")
-    fun getCraftsWithItemID(id: String): LiveData<List<Craft>>
+    fun getCraftsWithItemID(id: String): Flow<List<Craft>>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(craft: Craft)
