@@ -1,25 +1,36 @@
 package com.austinhodak.tarkovapi.room.dao
 
-import androidx.lifecycle.LiveData
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
+import androidx.room.*
+import com.austinhodak.tarkovapi.room.enums.ItemTypes
 import com.austinhodak.tarkovapi.room.models.Ammo
 import com.austinhodak.tarkovapi.room.models.Item
-import com.austinhodak.tarkovapi.room.models.PriceItem
+import com.austinhodak.tarkovapi.room.models.Pricing
 import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface ItemDao {
-    @Query("SELECT * FROM items ORDER BY name")
-    suspend fun getAll(): List<Item>
-
-    @Query("SELECT * FROM items ORDER BY name")
-    fun getAllLive(): LiveData<List<Item>>
 
     @Query("SELECT * FROM items WHERE id = :id")
-    fun getByID(id: String): Item
+    fun getByID(id: String): Flow<Item>
+
+    @Query("SELECT * FROM items WHERE id IN (:ids)")
+    suspend fun getByID(ids: List<String>): List<Item>
+
+    @Transaction
+    @Query("SELECT * FROM items WHERE itemType = :type")
+    fun getByType(type: ItemTypes): Flow<List<Item>>
+
+    @Transaction
+    @Query("SELECT * FROM items WHERE itemType in (:type)")
+    suspend fun getByTypes(type: List<ItemTypes>): List<Item>
+
+    @Transaction
+    @Query("SELECT * FROM items WHERE itemType in (:type)")
+    fun getByTypesArmor(type: List<ItemTypes>): Flow<List<Item>>
+
+    @Transaction
+    @Query("SELECT id, itemType, parent, Name, ShortName, pricing, Width, Height, BackgroundColor FROM items WHERE pricing IS NOT NULL")
+    fun getAllItems(): Flow<List<Item>>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(item: Item)
@@ -27,38 +38,22 @@ interface ItemDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insert(item: Ammo)
 
-    @Query("SELECT * FROM itemPrices")
-    fun getAllPricesFlow(): Flow<List<PriceItem>>
-
-    /*@Query("UPDATE items SET pricing = :tt WHERE id = :id")
-    suspend fun updateItemsTable(id: String, tt: ItemFragment)
+    @Query("UPDATE items SET pricing = :tt WHERE id = :id")
+    suspend fun updateItemsTable(id: String, tt: Pricing)
 
     @Query("UPDATE ammo SET pricing = :tt WHERE id = :id")
-    suspend fun updateAmmoTable(id: String, tt: ItemFragment)
+    suspend fun updateAmmoTable(id: String, tt: Pricing)
 
     @Query("UPDATE weapons SET pricing = :tt WHERE id = :id")
-    suspend fun updateWeaponTable(id: String, tt: ItemFragment)
+    suspend fun updateWeaponTable(id: String, tt: Pricing)
 
     @Transaction
-    suspend fun updateAllPricing(id: String?, pricing: ItemFragment) {
+    suspend fun updateAllPricing(id: String?, pricing: Pricing) {
         if (id != null) {
             updateItemsTable(id, pricing)
             updateAmmoTable(id, pricing)
             updateWeaponTable(id, pricing)
         }
-    }*/
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insert(item: PriceItem)
-
-    /*//TESTING
-    @Transaction
-    @Query("SELECT * FROM items")
-    fun getAllItemsWithPrices(): List<ItemWithPriceItem>
-
-    @Transaction
-    @Query("SELECT * FROM items WHERE id IS :id")
-    fun getAllItemsWithPrices(id: String): LiveData<ItemWithPriceItem>*/
-
+    }
 
 }
