@@ -1,25 +1,24 @@
 package com.austinhodak.tarkovapi.room.dao
 
-import androidx.lifecycle.LiveData
-import androidx.room.Dao
-import androidx.room.Insert
-import androidx.room.OnConflictStrategy
-import androidx.room.Query
-import com.austinhodak.tarkovapi.room.models.Ammo
+import androidx.room.*
 import com.austinhodak.tarkovapi.room.models.Weapon
+import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface WeaponDao {
 
-    @Query("SELECT * FROM ammo")
-    fun getAllAmmo(): LiveData<List<Ammo>>
+    @Transaction
+    @Query("SELECT * FROM weapons WHERE weapClass = :classID AND BackgroundColor IS NOT NULL AND Rarity IS NOT 'Not_exist' ORDER BY ShortName")
+    fun getWeaponsByClass(classID: String): Flow<List<Weapon>>
 
-    @Query("SELECT * FROM ammo WHERE id = :id")
-    fun getAmmoByID(id: String): Ammo
-
-    @Query("SELECT * FROM ammo WHERE Caliber = :caliber")
-    suspend fun getAmmoByCaliber(caliber: String): List<Ammo>
+    @Transaction
+    @Query("SELECT * FROM weapons WHERE id = :id")
+    fun getWeapon(id: String): Flow<Weapon>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insert(item: Weapon)
+    suspend fun insertAll(data: List<Weapon>)
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun insert(data: Weapon)
+
 }
