@@ -2,21 +2,19 @@ package com.austinhodak.tarkovapi.room.models
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import com.austinhodak.tarkovapi.fragment.TaskItem
-import kotlin.math.roundToInt
 
 @Entity(tableName = "barters")
 data class Barter(
-    @PrimaryKey val id: Int,
-    val source: String,
-    val requiredItems: List<TaskItem>?,
-    val rewardItems: List<TaskItem>?,
+    @PrimaryKey(autoGenerate = true) val id: Int? = null,
+    val requiredItems: List<Craft.CraftItem?>? = null,
+    val rewardItems: List<Craft.CraftItem?>? = null,
+    val source: String? = null
 ) {
     fun totalCost(): Int {
-        return requiredItems?.sumBy { (it.count * (it.item.fragments.itemFragment.avg24hPrice ?: 0)).roundToInt() } ?: 0
+        return requiredItems?.sumOf { (it?.count!!.times((it.item?.avg24hPrice ?: 0))) } ?: 0
     }
 
     fun estimatedProfit(): Int? {
-        return rewardItems?.first()?.item?.fragments?.itemFragment?.avg24hPrice?.minus(totalCost())
+        return rewardItems?.first()?.item?.avg24hPrice?.minus(totalCost())
     }
 }
